pyproject.toml
[project]
name = "langgraph-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.7",
    "langgraph>=0.5.1",
    "langsmith>=0.4.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
